Create LICENSE
This repository utilizes the GNU GPL v.2 license. This means the available code is available for public, private, commercial, and enterprise use but under a strongly protected license. 

If you wish to utilize pyTrader but are not certain if your use will violate any licensing, please email me at the email listed on my Github Profile with the header "pyTrader Permission" and I will gladly discuss it.